Tidy up capture
