+++ changedetectionio/content_fetchers/playwright.py
@@ -4,10 +4,101 @@ from urllib.parse import urlparse
 
 from loguru import logger
 
-from changedetectionio.content_fetchers import SCREENSHOT_MAX_HEIGHT_DEFAULT, visualselector_xpath_selectors
-from changedetectionio.content_fetchers.screenshot_handler import capture_full_page
+from changedetectionio.content_fetchers import SCREENSHOT_MAX_HEIGHT_DEFAULT, visualselector_xpath_selectors, \
+    SCREENSHOT_SIZE_STITCH_THRESHOLD, MAX_TOTAL_HEIGHT
+from changedetectionio.content_fetchers.screenshot_handler import stitch_images_worker
 from changedetectionio.content_fetchers.base import Fetcher, manage_user_agent
 from changedetectionio.content_fetchers.exceptions import PageUnloadable, Non200ErrorCodeReceived, EmptyReply, ScreenshotUnavailable
+
+
+
+def capture_full_page(page):
+    import os
+    import time
+    from multiprocessing import Process, Pipe
+
+    start = time.time()
+
+    page_height = page.evaluate("document.documentElement.scrollHeight")
+
+    logger.debug(f"Playwright viewport size {page.viewport_size}")
+
+    ############################################################
+    #### SCREENSHOT FITS INTO ONE SNAPSHOT (SMALLER PAGES) #####
+    ############################################################
+
+    # Optimization to avoid unnecessary stitching if we can avoid it
+    # Use the default screenshot method for smaller pages to take advantage
+    # of GPU and native playwright screenshot optimizations
+    # - No PIL needed here, no danger of memory leaks, no sub process required
+    if (page_height < SCREENSHOT_SIZE_STITCH_THRESHOLD and page_height < MAX_TOTAL_HEIGHT ):
+        logger.debug("Using default screenshot method")
+        page.request_gc()
+        screenshot = page.screenshot(
+            type="jpeg",
+            quality=int(os.getenv("SCREENSHOT_QUALITY", 30)),
+            full_page=True,
+        )
+        page.request_gc()
+        logger.debug(f"Screenshot captured in {time.time() - start:.2f}s")
+        return screenshot
+
+
+
+    ###################################################################################
+    #### CASE FOR LARGE SCREENSHOTS THAT NEED TO BE TRIMMED DUE TO MEMORY ISSUES  #####
+    ###################################################################################
+    # - PIL can easily allocate memory and not release it cleanly
+    # - Fetching screenshot from playwright seems  OK
+    # Image.new is leaky even with .close()
+    # So lets prepare all the data chunks and farm it out to a subprocess for clean memory handling
+
+    logger.debug(
+        "Using stitching method for large screenshot because page height exceeds threshold"
+    )
+
+    # Limit the total capture height
+    capture_height = min(page_height, MAX_TOTAL_HEIGHT)
+
+    # Calculate number of chunks needed using ORIGINAL viewport height
+    num_chunks = (capture_height + page.viewport_size['height'] - 1) // page.viewport_size['height']
+    screenshot_chunks = []
+
+    # Track cumulative paste position
+    y_offset = 0
+    for _ in range(num_chunks):
+
+        page.request_gc()
+        page.evaluate(f"window.scrollTo(0, {y_offset})")
+        page.request_gc()
+        h = min(page.viewport_size['height'], capture_height - y_offset)
+        screenshot_chunks.append(page.screenshot(
+                type="jpeg",
+                clip={
+                    "x": 0,
+                    "y": 0,
+                    "width": page.viewport_size['width'],
+                    "height": h,
+                },
+                quality=int(os.getenv("SCREENSHOT_QUALITY", 30)),
+            ))
+
+        y_offset += h # maybe better to inspect the image here?
+        page.request_gc()
+
+    # PIL can leak memory in various situations, assign the work to a subprocess for totally clean handling
+
+    parent_conn, child_conn = Pipe()
+    p = Process(target=stitch_images_worker, args=(child_conn, screenshot_chunks, page_height, capture_height))
+    p.start()
+    result = parent_conn.recv_bytes()
+    p.join()
+
+    screenshot_chunks = None
+    logger.debug(f"Screenshot - Page height: {page_height} Capture height: {capture_height} - Stitched together in {time.time() - start:.2f}s")
+
+    return result
+
 
 class fetcher(Fetcher):
     fetcher_description = "Playwright {}/Javascript".format(
@@ -61,7 +152,9 @@ class fetcher(Fetcher):
 
     def screenshot_step(self, step_n=''):
         super().screenshot_step(step_n=step_n)
-        screenshot = capture_full_page(self.page)
+        original_viewport = self.page.viewport_size
+        screenshot = capture_full_page(page=self.page, viewport_width=original_viewport["width"], viewport_height=original_viewport["height"])
+
 
         if self.browser_steps_screenshot_path is not None:
             destination = os.path.join(self.browser_steps_screenshot_path, 'step_{}.jpeg'.format(step_n))
@@ -164,7 +257,11 @@ class fetcher(Fetcher):
                 raise PageUnloadable(url=url, status_code=None, message=str(e))
 
             if self.status_code != 200 and not ignore_status_codes:
-                screenshot = capture_full_page(self.page)
+                original_viewport = self.page.viewport_size
+                screenshot = capture_full_page(page=self.page,
+                                               viewport_width=original_viewport["width"],
+                                               viewport_height=original_viewport["height"])
+
                 raise Non200ErrorCodeReceived(url=url, status_code=self.status_code, screenshot=screenshot)
 
             if not empty_pages_are_a_change and len(self.page.content().strip()) == 0:
@@ -211,7 +308,7 @@ class fetcher(Fetcher):
             # acceptable screenshot quality here
             try:
                 # The actual screenshot - this always base64 and needs decoding! horrible! huge CPU usage
-                self.screenshot = capture_full_page(self.page)
+                self.screenshot = capture_full_page(page=self.page)
 
             except Exception as e:
                 # It's likely the screenshot was too long/big and something crashed
