+++ changedetectionio/content_fetchers/puppeteer.py
@@ -6,9 +6,93 @@ from urllib.parse import urlparse
 
 from loguru import logger
 
-from changedetectionio.content_fetchers import SCREENSHOT_MAX_HEIGHT_DEFAULT, visualselector_xpath_selectors
+from changedetectionio.content_fetchers import SCREENSHOT_MAX_HEIGHT_DEFAULT, visualselector_xpath_selectors, \
+    SCREENSHOT_SIZE_STITCH_THRESHOLD, MAX_TOTAL_HEIGHT
 from changedetectionio.content_fetchers.base import Fetcher, manage_user_agent
 from changedetectionio.content_fetchers.exceptions import PageUnloadable, Non200ErrorCodeReceived, EmptyReply, BrowserFetchTimedOut, BrowserConnectError
+from changedetectionio.content_fetchers.screenshot_handler import stitch_images_worker
+
+
+# Bug 3 in Playwright screenshot handling
+# Some bug where it gives the wrong screenshot size, but making a request with the clip set first seems to solve it
+
+# Screenshots also travel via the ws:// (websocket) meaning that the binary data is base64 encoded
+# which will significantly increase the IO size between the server and client, it's recommended to use the lowest
+# acceptable screenshot quality here
+async def capture_full_page(page):
+    import os
+    import time
+    from multiprocessing import Process, Pipe
+
+    start = time.time()
+
+    page_height = await page.evaluate("document.documentElement.scrollHeight")
+
+    logger.debug(f"Puppeteer viewport size {page.viewport}")
+
+    ############################################################
+    #### SCREENSHOT FITS INTO ONE SNAPSHOT (SMALLER PAGES) #####
+    ############################################################
+
+    # Optimization to avoid unnecessary stitching if we can avoid it
+    # Use the default screenshot method for smaller pages to take advantage
+    # of GPU and native playwright screenshot optimizations
+    # - No PIL needed here, no danger of memory leaks, no sub process required
+    if (page_height < SCREENSHOT_SIZE_STITCH_THRESHOLD and page_height < MAX_TOTAL_HEIGHT ):
+        logger.debug("Using default screenshot method")
+        await page.evaluate(f"window.scrollTo(0, 0)")
+        screenshot = await page.screenshot(
+            type_="jpeg",
+            quality=int(os.getenv("SCREENSHOT_QUALITY", 30)),
+            fullPage=True,
+        )
+        logger.debug(f"Screenshot captured in {time.time() - start:.2f}s")
+        return screenshot
+
+    ###################################################################################
+    #### CASE FOR LARGE SCREENSHOTS THAT NEED TO BE TRIMMED DUE TO MEMORY ISSUES  #####
+    ###################################################################################
+    # - PIL can easily allocate memory and not release it cleanly
+    # - Fetching screenshot from playwright seems  OK
+    # Image.new is leaky even with .close()
+    # So lets prepare all the data chunks and farm it out to a subprocess for clean memory handling
+
+    logger.debug(
+        "Using stitching method for large screenshot because page height exceeds threshold"
+    )
+
+    # Limit the total capture height
+    capture_height = min(page_height, MAX_TOTAL_HEIGHT)
+
+    # Calculate number of chunks needed using ORIGINAL viewport height
+    num_chunks = (capture_height + page.viewport['height'] - 1) // page.viewport['height']
+    screenshot_chunks = []
+
+    # Track cumulative paste position
+    y_offset = 0
+    for _ in range(num_chunks):
+        await page.evaluate(f"window.scrollTo(0, {y_offset})")
+        h = min(page.viewport['height'], capture_height - y_offset)
+        screenshot_chunks.append(await page.screenshot(
+                type_="jpeg",
+                quality=int(os.getenv("SCREENSHOT_QUALITY", 30)),
+            ))
+
+        y_offset += h # maybe better to inspect the image here?
+
+    # PIL can leak memory in various situations, assign the work to a subprocess for totally clean handling
+
+    parent_conn, child_conn = Pipe()
+    p = Process(target=stitch_images_worker, args=(child_conn, screenshot_chunks, page_height, capture_height))
+    p.start()
+    result = parent_conn.recv_bytes()
+    p.join()
+
+    screenshot_chunks = None
+    logger.debug(f"Screenshot - Page height: {page_height} Capture height: {capture_height} - Stitched together in {time.time() - start:.2f}s")
+
+    return result
+
 
 class fetcher(Fetcher):
     fetcher_description = "Puppeteer/direct {}/Javascript".format(
@@ -181,11 +265,10 @@ class fetcher(Fetcher):
             raise PageUnloadable(url=url, status_code=None, message=str(e))
 
         if self.status_code != 200 and not ignore_status_codes:
-            screenshot = await self.page.screenshot(type_='jpeg',
-                                                    fullPage=True,
-                                                    quality=int(os.getenv("SCREENSHOT_QUALITY", 30)))
+            screenshot = await capture_full_page(page=self.page)
 
             raise Non200ErrorCodeReceived(url=url, status_code=self.status_code, screenshot=screenshot)
+
         content = await self.page.content
 
         if not empty_pages_are_a_change and len(content.strip()) == 0:
@@ -203,7 +286,7 @@ class fetcher(Fetcher):
 
         # So we can find an element on the page where its selector was entered manually (maybe not xPath etc)
         # Setup the xPath/VisualSelector scraper
-        if current_include_filters is not None:
+        if current_include_filters:
             js = json.dumps(current_include_filters)
             await self.page.evaluate(f"var include_filters={js}")
         else:
@@ -214,40 +297,20 @@ class fetcher(Fetcher):
             "visualselector_xpath_selectors": visualselector_xpath_selectors,
             "max_height": MAX_TOTAL_HEIGHT
         })
+        if not self.xpath_data:
+            raise Exception(f"Content Fetcher > xPath scraper failed. Please report this URL so we can fix it :)")
 
         self.instock_data = await self.page.evaluate(self.instock_data_js)
 
         self.content = await self.page.content
-        # Bug 3 in Playwright screenshot handling
-        # Some bug where it gives the wrong screenshot size, but making a request with the clip set first seems to solve it
-        # JPEG is better here because the screenshots can be very very large
 
-        # Screenshots also travel via the ws:// (websocket) meaning that the binary data is base64 encoded
-        # which will significantly increase the IO size between the server and client, it's recommended to use the lowest
-        # acceptable screenshot quality here
-        #@todo use the screenshot_handler but it must be used as some kind of async process
-        try:
-            self.screenshot = await self.page.screenshot(type_='jpeg',
-                                                         fullPage=True,
-                                                         quality=int(os.getenv("SCREENSHOT_QUALITY", 30)))
-        except Exception as e:
-            logger.error("Error fetching screenshot")
-            # // May fail on very large pages with 'WARNING: tile memory limits exceeded, some content may not draw'
-            # // @ todo after text extract, we can place some overlay text with red background to say 'croppped'
-            logger.error('ERROR: content-fetcher page was maybe too large for a screenshot, reverting to viewport only screenshot')
-            try:
-                self.screenshot = await self.page.screenshot(type_='jpeg',
-                                                             fullPage=False,
-                                                             quality=int(os.getenv("SCREENSHOT_QUALITY", 30)))
-            except Exception as e:
-                logger.error('ERROR: Failed to get viewport-only reduced screenshot :(')
-                pass
-        finally:
-            # It's good to log here in the case that the browser crashes on shutting down but we still get the data we need
-            logger.success(f"Fetching '{url}' complete, closing page")
-            await self.page.close()
-            logger.success(f"Fetching '{url}' complete, closing browser")
-            await browser.close()
+        self.screenshot = await capture_full_page(page=self.page)
+
+        # It's good to log here in the case that the browser crashes on shutting down but we still get the data we need
+        logger.success(f"Fetching '{url}' complete, closing page")
+        await self.page.close()
+        logger.success(f"Fetching '{url}' complete, closing browser")
+        await browser.close()
         logger.success(f"Fetching '{url}' complete, exiting puppeteer fetch.")
 
     async def main(self, **kwargs):
